requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
face-recognition==1.3.0
dlib==19.24.2
Cython>=0.29.0
numpy>=1.21.0
pandas>=1.5.0
pillow>=9.0.0
requests>=2.28.0
transformers>=4.20.0
torch>=1.12.0
scikit-learn>=1.1.0
scipy>=1.7.0
plotly>=5.10.0
python-dotenv>=0.19.0
schedule>=1.1.0
setuptools>=65.0.0
wheel>=0.37.0
